[components] For https://github.com/mozilla-mobile/android-components/issues/11954 - Adapt crash prompt for background child process crashes
Although we currently prefer showing the crash notification over the
crash prompt for background process crashes, users can still open the
prompt by clicking on the notification. Unfortunately in its current
state the prompt is unsuitable for background process crashes. This
patch makes some adjustments to the prompt for the background process
case to make it less jarring:

* The title now matches the crash notification title for background
  child process crashes, eg "Sorry. A problem occurred in Firefox."
  rather than "Sorry. Firefox had a problem and crashed."
* We avoid setting the FLAG_ACTIVITY_CLEAR_TASK flag on the intent,
  meaning the browser remains visible behind the dialog.
* We remove the "Restart Firefox" button, as the browser has continued
  running and therefore does not need to be restarted.