[fenix] For https://github.com/mozilla-mobile/fenix/issues/25401 - Improve how sponsored stories impressions are reported
While testing I observed that the Rect returned from `boundsInWindow()`
- will have the full height and width of the composable even if scrolled
offscreen vertically
- will only have the height and width of what part of the composable is shown
pn the screen.
Possibly because of the interactions of a vertically scrollable RecyclerView
and the LazyRow the composable is shown in.

To account for this the method calculating how much of the composable is shown
on screen will also now receive the real composable size as an argument.