[fenix] For https://github.com/mozilla-mobile/fenix/issues/25538 - Add the "Wallpapers" tag to the new telemetry
By using a feature tag we can avoid having to define new tags in the Glean
annotations repository.

The list of available tags needed to be regenerated to be able to reference the
new "Wallpapers" tag.
Because the tags regeneration also brought a change in tags from "TopSites" to
"Shortcuts" all previous telemetry using "TopSites" had to be migrated to use
the new "Shortcuts" tag.