[components] For https://github.com/mozilla-mobile/android-components/issues/12433 - Separate responsibilities regarding the media notification
This will avoid overlapping responsibilities for when and which media
notification to show by ensuring that:
- MediaSessionFeature is the only one observing the store for media updates and
deciding which notification to show by dispatching updates to the media service
- MediaSessionServiceDelegate is responsible to act upon calls from
MediaSessionFeature to show/hide/update a media notification and upon calls
from the media notification action button to resume or play media.

These changes also ensure we'll not get errors about not starting as foreground
by using a bound service that will call start foreground to show the media
notification only after that is build and available.