[fenix] For https://github.com/mozilla-mobile/fenix/issues/26555 - Observe and update the wallpaper before HomeScreen is visible.
By using Store.observeManually in a standalone coroutine we can observe the
store and update the wallpapers even before onStart (in manual tests is right
around onStart, certainly before the other widgets shown on homescreen).

Created a new WallpapersObserver to have the functionality easier to reason
about and be easier to test.