Bug 1796658 - fix notifications on failure of geckoview nightly bumps
A few years ago the owner and branch names for these bumps changed, so
the notification would never actually happen.
Now we look for:
- the complete-pr task
- on the relbot/upgrade-geckoview-ac-main branch of the base repository

Since the branch name no longer includes the geckoview build id, we no
longer include it in the email subject and replace it with the PR
number.