[fenix] For https://github.com/mozilla-mobile/fenix/issues/25816: Ensure the "show search suggestions" user option is followed
Previously the check for the "Search -> Show search suggestions" user setting
was only used in the default SearchFragmentState but not again if users change
the current search engine as part of the unified search feature.
This comes to ensure that that check is always made when needing to configure
new search engine results.