Bug 1811928 - Android Components: get metrics/pings.yaml for crash pings from mozilla-central
It was determined that the best approach would be to manually copy the
files. Once android-components is merged into mozilla-central the files
can be potentially easily shared.

The use of kapt is replaced by ksp as the kapt produces java stubs from
the glean generated code which are invalid java (ksp does not generate
stubs).