Bug 1816196 - Do not fetch messaging during storage initialization
While debugging a UI test, we found that the messaging feature was being
queried during the storage initialization and acting on the cached value
alone. This could lead us to no longer receiving new messages during the
app lifetime.

For the unit tests, we switched from mocking the entire feature to
mocking just the `FeaturesInterface` as the test was providing
false-positive results. It's better to create a real feature in any case
for the other testing needs.

Co-authored-by: Benjamin Forehand Jr <bennyjr169@gmail.com>