Bug 1817944 - Add nightly official flag to both apps
After the migration to the monorepo, the `include-nightly-version` flag
seems to have gotten lost.

The effect of this flag is that builds are not marked as "official" and
features that only work with this (WebAuthn) are broken as a result.

This doesn't affect Focus since the WebAuthn feature doesn't exist in
that app and there a no other usages for the official flag either.

Co-authored-by: mcarare <mihai.carare.dev@gmail.com>
Co-authored-by: Christian Sadilek <christian.sadilek@gmail.com>